+++ src/cli.js
@@ -568,10 +568,20 @@ async function build() {
 
     tailwindPlugin.postcss = true
 
+    let IMPORT_COMMENT = '__TAILWIND_RESTORE_IMPORT__: '
+
     let [beforePlugins, afterPlugins, postcssOptions] = includePostCss
       ? await loadPostCssPlugins()
       : [
           [
+            (root) => {
+              root.walkAtRules('import', (rule) => {
+                if (rule.params.slice(1).startsWith('tailwindcss/')) {
+                  rule.after(postcss.comment({ text: IMPORT_COMMENT + rule.params }))
+                  rule.remove()
+                }
+              })
+            },
             (() => {
               try {
                 return require('postcss-import')
@@ -579,6 +589,19 @@ async function build() {
 
               return lazyPostcssImport()
             })(),
+            (root) => {
+              root.walkComments((rule) => {
+                if (rule.text.startsWith(IMPORT_COMMENT)) {
+                  rule.after(
+                    postcss.atRule({
+                      name: 'import',
+                      params: rule.text.replace(IMPORT_COMMENT, ''),
+                    })
+                  )
+                  rule.remove()
+                }
+              })
+            },
           ],
           [],
           {},